pyproject.toml
[tool.poetry]
name = "protomate"
version = "0.6.0"
description = "Python built CLI tool for automated github project initialization."
authors = ["Redowan Delowar"]
license = "MIT"

readme = "README.md"
repository = "https://github.com/rednafi/protomate"
documentation = "https://rednafi.github.io/protomate/"

classifiers=[
        "License :: OSI Approved :: MIT License",
        "Programming Language :: Python",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.7",
        "Programming Language :: Python :: Implementation :: CPython",
        "Programming Language :: Python :: Implementation :: PyPy"
    ]

[tool.poetry.dependencies]
python = "^3.6"
art="4.0"
colorama="0.4.1"
PyGithub="1.43.8"
keyring="17.1.1"
loguru="0.3.2"
prompt_toolkit="2.0.9"
questionary="1.3.0"
sty= "1.0.0b12"
termcolor="1.1.0"

[tool.poetry.dev-dependencies]
setuptools = "41.2.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
flake8 = "^3.7"
[tool.poetry.scripts]
protomate = 'protomate.script:main'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

